# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação da API",
        "description": "Petship API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "https://api.petship.pt",
            "description": "Petship API"
        }
    ],
    "paths": {
...
        "/api/publicaranimal": {
            "post": {
                "tags": [
                    "Animais"
                ],
                "summary": "Registar um animal",
                "description": "",
                "operationId": "2527b12868b5ffa6f9e590839cf1c9cd",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data_recolha",
                                    "ferido",
                                    "agressivo",
                                    "nome",
                                    "sexo",
                                    "especie",
                                    "raca",
                                    "porte",
                                    "idade",
                                    "cor"
                                ],
                                "properties": {
                                    "data_recolha": {
                                        "example": "Coimbra"
                                    },
                                    "ferido": {
                                        "example": "1"
                                    },
                                    "agressivo": {
                                        "example": "Coimbra"
                                    },
                                    "nome": {
                                        "example": "Rufus"
                                    },
                                    "sexo": {
                                        "example": "1"
                                    },
                                    "especie": {
                                        "example": "1"
                                    },
                                    "raca": {
                                        "example": "1"
                                    },
                                    "porte": {
                                        "example": "1"
                                    },
                                    "idade": {
                                        "example": "1"
                                    },
                                    "cor": {
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/associacao/animais": {
            "get": {
                "tags": [
                    "Animais"
                ],
                "summary": "Listar animais da associação logada",
                "description": "",
                "operationId": "b7375e8ad240d92bcf3821c06ba8282f",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/removeranimal/{id}": {
            "post": {
                "tags": [
                    "Animais"
                ],
                "summary": "Remover um animal",
                "description": "",
                "operationId": "c1c76b39591f15e463ae875981bccf89",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do animal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/associacao/animal/num/{id}": {
            "get": {
                "tags": [
                    "Animais"
                ],
                "summary": "Obter dados numéricos de um animal",
                "description": "",
                "operationId": "a96a3aa223ce004b7a40ffe888f77b87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do animal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/editaranimal/{id}": {
            "post": {
                "tags": [
                    "Animais"
                ],
                "summary": "Editar dados de um animal",
                "description": "",
                "operationId": "35405237efce3ac032e01676407748f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do animal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data_recolha",
                                    "ferido",
                                    "agressivo",
                                    "nome",
                                    "sexo",
                                    "especie",
                                    "raca",
                                    "porte",
                                    "idade",
                                    "cor"
                                ],
                                "properties": {
                                    "data_recolha": {
                                        "example": "Coimbra"
                                    },
                                    "ferido": {
                                        "example": "1"
                                    },
                                    "agressivo": {
                                        "example": "Coimbra"
                                    },
                                    "nome": {
                                        "example": "Rufus"
                                    },
                                    "sexo": {
                                        "example": "1"
                                    },
                                    "especie": {
                                        "example": "1"
                                    },
                                    "raca": {
                                        "example": "1"
                                    },
                                    "porte": {
                                        "example": "1"
                                    },
                                    "idade": {
                                        "example": "1"
                                    },
                                    "cor": {
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "token": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}